refactor into `tileset.py`
